fix autosar

+++ mc2/inplace_matmul_all_reduce_add_rms_norm/examples/test_aclnn_inplace_quant_matmul_all_reduce_add_rms_norm.cpp
@@ -154,7 +154,7 @@ int launchOneThreadQuantMatmulAllReduceAddRmsNorm(Args &args) {
     CHECK_RET(ret == ACL_SUCCESS,
             LOG_PRINT("aclnnInplaceQuantMatmulAllReduceAddRmsNormGetWorkspaceSize failed. ERROR: %d\n", ret); return ret);
     // 根据第一段接口计算出的workspaceSize申请device内存
-    if (workspaceSize > 0) {
+    if (workspaceSize > static_cast<uint64_t>(0)) {
         ret = aclrtMalloc(&workspaceAddr, workspaceSize, ACL_MEM_MALLOC_HUGE_FIRST);
         CHECK_RET(ret == ACL_SUCCESS, LOG_PRINT("allocate workspace failed. ERROR: %d\n", ret); return ret);
     }
@@ -216,7 +216,7 @@ int launchOneThreadQuantMatmulAllReduceAddRmsNorm(Args &args) {
     if (normOutDeviceAddr != nullptr) {
         aclrtFree(normOutDeviceAddr);
     }
-    if (workspaceSize > 0) {
+    if (workspaceSize > static_cast<uint64_t>(0)) {
         aclrtFree(workspaceAddr);
     }
     aclrtDestroyStream(args.stream);
@@ -246,7 +246,7 @@ int main(int argc, char *argv[]) {
     Args args[ndev];
     aclrtStream stream[ndev];
     aclrtContext context[ndev];
-    for (uint32_t rankId = 0; rankId < ndev; rankId++) {
+    for (uint32_t rankId = 0; rankId < static_cast<uint32_t>(ndev); rankId++) {
         ret = aclrtSetDevice(rankId);
         CHECK_RET(ret == ACL_SUCCESS, LOG_PRINT("aclrtSetDevice failed. ERROR: %d\n", ret); return ret);
         ret = aclrtCreateContext(&context[rankId], rankId);
@@ -256,7 +256,7 @@ int main(int argc, char *argv[]) {
     }
     // 启动多线程
     std::vector<std::unique_ptr<std::thread>> threads(ndev);
-    for (uint32_t rankId = 0; rankId < ndev; rankId++) {
+    for (uint32_t rankId = 0; rankId < static_cast<uint32_t>(ndev); rankId++) {
         args[rankId].rankId = rankId;
         args[rankId].hcclComm = comms[rankId];
         args[rankId].stream = stream[rankId];
@@ -264,7 +264,7 @@ int main(int argc, char *argv[]) {
         threads[rankId].reset(
                 new(std::nothrow) std::thread(&launchOneThreadQuantMatmulAllReduceAddRmsNorm, std::ref(args[rankId])));
     }
-    for (uint32_t rankId = 0; rankId < ndev; rankId++) {
+    for (uint32_t rankId = 0; rankId < static_cast<uint32_t>(ndev); rankId++) {
         threads[rankId]->join();
     }
     aclFinalize();
